[TASK] Move Event Listener Test
Due to limitations and better wording
Event listener is moved to separate functional
test, as this makes more sense and shows
maintainer and contributor, what is tested here.
